+++ lib/src/models/facebook_access_token.dart
@@ -6,20 +6,23 @@ class FacebookAccessToken {
   final String userId;
   final DateTime expires;
   final List<String> permissions;
+  final List<String> declinedPermissions;
 
   FacebookAccessToken.fromMap(Map<String, dynamic> map)
       : token = map['token'],
         userId = map['userId'],
         expires =
             DateTime.fromMillisecondsSinceEpoch(map['expires'], isUtc: true),
-        permissions = map['permissions'].cast<String>();
+        permissions = map['permissions'].cast<String>(),
+        declinedPermissions = map['declinedPermissions'].cast<String>();
 
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'token': token,
       'userId': userId,
       'expires': expires.millisecondsSinceEpoch,
-      'permissions': permissions
+      'permissions': permissions,
+      'declinedPermissions': declinedPermissions,
     };
   }
 
@@ -31,18 +34,21 @@ class FacebookAccessToken {
           token == other.token &&
           userId == other.userId &&
           expires == other.expires &&
-          permissions.isUnorderedEquivalent(other.permissions);
+          permissions.isUnorderedEquivalent(other.permissions) &&
+          declinedPermissions.isUnorderedEquivalent(other.declinedPermissions);
 
   @override
   int get hashCode =>
       token.hashCode ^
       userId.hashCode ^
       expires.hashCode ^
-      permissions.hashCode;
+      permissions.hashCode ^
+      declinedPermissions.hashCode;
 
   @override
   String toString() {
     return 'FacebookAccessToken(token: $token, userId: $userId, '
-        'expires: $expires, permissions: $permissions)';
+        'expires: $expires, permissions: $permissions, '
+        'declinedPermissions: $declinedPermissions)';
   }
 }