+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next release
+
+* Add `declinedPermissions` in `AccessToken`.
+
 ## 0.1.1 - 2020-05-22
 
 * Add error information for login failure.